chore: Folders renamed to snake case
