+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "yokaidex",
-    "version": "0.1.0",
+    "version": "0.1.1",
     "description": "A PWA with the info from Yokai watch",
     "license": "MIT",
     "homepage": "https://joaopedrodcf.github.io/yokaidex",
