+++ src/components/main/style.js
